+++ jsonargparse_tests/test_signatures.py
@@ -1,3 +1,6 @@
+from __future__ import annotations
+
+import sys
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple, Union
 from unittest.mock import patch
@@ -85,7 +88,7 @@ def test_add_class_failure_not_a_class(parser):
 def test_add_class_failure_positional_without_type(parser):
     with pytest.raises(ValueError) as ctx:
         parser.add_class_arguments(Class2)
-    ctx.match(f'Parameter "c2_a0" from "{__name__}.Class2.__init__" does not specify a type')
+    ctx.match(f"Parameter 'c2_a0' from '{__name__}.Class2.__init__' does not specify a type")
 
 
 def test_add_class_without_nesting(parser):
@@ -251,6 +254,7 @@ def test_add_class_with_required_parameters(parser):
     assert cfg.model == Namespace(m=0.1, n=3)
 
 
+@pytest.mark.skipif(sys.version_info[:2] == (3, 6), reason="import forces future annotations in python 3.6")
 def test_add_class_conditional_kwargs(parser):
     from jsonargparse_tests.test_parameter_resolvers import ClassG
 
@@ -510,6 +514,16 @@ def test_add_function_implicit_optional(parser):
     assert None is parser.parse_args(["--a1=null"]).a1
 
 
+def func_type_as_string(a2: "int"):
+    return a2
+
+
+@pytest.mark.skipif(sys.version_info[:2] == (3, 6), reason="not supported in python 3.6")
+def test_add_function_fail_untyped_true_str_type(parser):
+    added_args = parser.add_function_arguments(func_type_as_string, fail_untyped=True)
+    assert ["a2"] == added_args
+
+
 def func_untyped_params(a1, a2=None):
     return a1
 
@@ -517,17 +531,7 @@ def func_untyped_params(a1, a2=None):
 def test_add_function_fail_untyped_true_untyped_params(parser):
     with pytest.raises(ValueError) as ctx:
         parser.add_function_arguments(func_untyped_params, fail_untyped=True)
-    ctx.match('Parameter "a1" from .* does not specify a type')
-
-
-def func_type_as_string(a2: "int"):
-    return a2
-
-
-def test_add_function_fail_untyped_true_str_type(parser):
-    with pytest.raises(ValueError) as ctx:
-        parser.add_function_arguments(func_type_as_string, fail_untyped=True)
-    ctx.match('Parameter "a2" from .* specifies the type as a string')
+    ctx.match("Parameter 'a1' from .* does not specify a type")
 
 
 def test_add_function_fail_untyped_false(parser):