+++ jsonargparse_tests/test_subclasses.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import json
 import textwrap
 import warnings